+++ packages/workflow/CHANGELOG.md
@@ -1,5 +1,39 @@
 # workflow
 
+## 5.0.0-beta.46
+
+### Patch Changes
+
+- Updated dependencies [[`0750fe8`](https://github.com/vercel/workflow/commit/0750fe8958f50ee354cd24fe3f02a37c801bc4dc), [`d9e0777`](https://github.com/vercel/workflow/commit/d9e0777eb8b1ce5f3be3fe865bc5a17fdbdb9d5d), [`d9e0777`](https://github.com/vercel/workflow/commit/d9e0777eb8b1ce5f3be3fe865bc5a17fdbdb9d5d), [`82e2678`](https://github.com/vercel/workflow/commit/82e267893917e06f5d3ce316baefc27cd024a912), [`556f3f0`](https://github.com/vercel/workflow/commit/556f3f080a33f8673cf3c08805997dba4e4d865e)]:
+  - @workflow/astro@5.0.0-beta.46
+  - @workflow/sveltekit@5.0.0-beta.46
+  - @workflow/core@5.0.0-beta.46
+  - @workflow/cli@5.0.0-beta.46
+  - @workflow/nest@5.0.0-beta.46
+  - @workflow/next@5.0.0-beta.46
+  - @workflow/nitro@5.0.0-beta.46
+  - @workflow/rollup@5.0.0-beta.46
+  - @workflow/typescript-plugin@5.0.0-beta.5
+  - @workflow/errors@5.0.0-beta.18
+  - @workflow/nuxt@5.0.0-beta.46
+
+## 5.0.0-beta.45
+
+### Patch Changes
+
+- Updated dependencies [[`d62b444`](https://github.com/vercel/workflow/commit/d62b44473b43e183e71386fe84b33f5e7bb5445c), [`27cab14`](https://github.com/vercel/workflow/commit/27cab14adcc6f748500fca19cf78feeb60a125e7), [`f7fb012`](https://github.com/vercel/workflow/commit/f7fb0126528fac2168492c3143fe0379b6f5ef1a), [`5841558`](https://github.com/vercel/workflow/commit/584155897f75e712a1c2bc199d6d12027cd18dab)]:
+  - @workflow/core@5.0.0-beta.45
+  - @workflow/astro@5.0.0-beta.45
+  - @workflow/cli@5.0.0-beta.45
+  - @workflow/nest@5.0.0-beta.45
+  - @workflow/next@5.0.0-beta.45
+  - @workflow/nitro@5.0.0-beta.45
+  - @workflow/rollup@5.0.0-beta.45
+  - @workflow/sveltekit@5.0.0-beta.45
+  - @workflow/typescript-plugin@5.0.0-beta.5
+  - @workflow/errors@5.0.0-beta.18
+  - @workflow/nuxt@5.0.0-beta.45
+
 ## 5.0.0-beta.44
 
 ### Minor Changes